hotel-booking-a2a-agent/llm-client/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "llm-client"
version = "0.1.0"
description = "LLM client for natural language processing with OpenAI-compatible APIs"
requires-python = ">=3.12"
dependencies = [
  "openai",
  "python-dotenv",
]

[project.scripts]
llm-client = "llm_client.main:main"

[dependency-groups]
dev = [
  "pytest",
  "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["llm-client"]
